% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_usgov}
\alias{list_usgov}
\title{List of all branches, departments, agencies, etc.}
\usage{
list_usgov()
}
\value{
A data.frame of rows matching the query
}
\description{
List of all branches, departments, agencies, etc.
}
\examples{
list_usgov()
}

